requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.9
jinja2==3.1.2
python-multipart==0.0.6

